pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "knowledgeforge"
version = "0.1.0"
description = "AI-powered knowledge graph builder and RAG system for the OpenClaw/NemoClaw ecosystem"
readme = "README.md"
license = "MIT"
requires-python = ">=3.12"

dependencies = [
    "mcp>=1.0",
    "httpx>=0.27",
    "google-genai>=1.0",
    "neo4j>=5.0",
    "networkx>=3.0",
    "numpy>=1.26",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "fastapi>=0.110",
    "uvicorn[standard]>=0.30",
    "apscheduler>=3.10",
    "python-dotenv>=1.0",
    "structlog>=24.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=5.0",
    "ruff>=0.4",
    "httpx",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
